+++ src/types.ts
@@ -136,6 +136,12 @@ export type MastodonApiObject = (
     mastodon.v1.TrendLink
 );
 
+export type MastoApiObject = (
+    Account |
+    MastodonApiObject |
+    Toot
+);
+
 // All these types have an id property
 export type MastodonObjWithID = (
     SerializableToot |