+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Add a new `FavouritedTagsScorer` (and a new `UserData.favouritedTagCounts` property)
 * Fix bug where new scorer weights were not automatically getting set to defaults
 * Remove trending tags from the list of participated tags we're pulling toots for because we get them anyways
+* Fix bug where home timeline toot cache would just grow (dedupe on the way in)
+* Add DEFAULT_REQUESTS configuration
 
 # v0.34.0
 * Use a value for the user displayed TIME_DECAY param that is 10x the value used in the calculation to make it more user friendly.